requirements.txt
django==1.11.1
django-compressor==2.1.1
django-extensions==1.7.9
django-htmlmin==0.10.0
django-jet==1.0.6
django-ordered-model=1.4.1
django-sass-processor==0.5.4
libsass==0.12.3
psycopg2==2.7.1
